# service-sdk/12.1.0/migrate-to-open-api/messaging-api/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: No description available
  title: Message Service API Spec
  version: v1
servers:
- description: The server
  url: /v1
tags:
- name: message service api spec
paths:
  /message:
    delete:
      description: Deletes a message
      operationId: deleteMessage
      parameters:
      - description: Id of the message record
        example: a5b0fe7d-c4dd-40a7-bd80-dfc7869327e1
        in: query
        name: id
        required: true
        schema:
          type: string
      responses:
        204:
          description: request to delete completed
        400:
          content:
            application/json:
              example:
                message: Bad Request
                errors:
                - message: Value Exceeded. Must be between {min} and {max}.
                  key: common.api.shoesize
                  context:
                    max: "50"
                    min: "1"
              schema:
                $ref: '#/components/schemas/BadRequestError'
          description: BadRequest
        500:
          content:
            application/json:
              example:
                message: Description of error
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: InternalServerError
      summary: Adds or updates a message.
      tags:
      - message
    description: No description available
    get:
      description: Returns available items, optionally filtered by name
      operationId: getMessage
      parameters:
      - description: Id of the message record
        example: a5b0fe7d-c4dd-40a7-bd80-dfc7869327e1
        in: query
        name: id
        required: true
        schema:
          type: string
      responses:
        200:
          content:
            application/json:
              example:
                $ref: examples/body/example-message.json
              schema:
                $ref: '#/components/schemas/Message'
          description: No description available
        400:
          content:
            application/json:
              example:
                message: Bad Request
                errors:
                - message: Value Exceeded. Must be between {min} and {max}.
                  key: common.api.shoesize
                  context:
                    max: "50"
                    min: "1"
              schema:
                $ref: '#/components/schemas/BadRequestError'
          description: BadRequest
        500:
          content:
            application/json:
              example:
                message: Description of error
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: InternalServerError
      summary: Returns available items, optionally filtered by name.
      tags:
      - message
    post:
      description: Adds or updates a message
      operationId: postMessage
      requestBody:
        content:
          application/json:
            example:
              $ref: examples/body/example-message.json
            schema:
              $ref: '#/components/schemas/Message'
        description: Adds or updates a message
      responses:
        201:
          content:
            application/json:
              example:
                $ref: examples/body/example-message-created.json
              schema:
                $ref: '#/components/schemas/MessagePostResponseBody'
          description: request to create item accepted
        400:
          content:
            application/json:
              example:
                message: Bad Request
                errors:
                - message: Value Exceeded. Must be between {min} and {max}.
                  key: common.api.shoesize
                  context:
                    max: "50"
                    min: "1"
              schema:
                $ref: '#/components/schemas/BadRequestError'
          description: BadRequest
        500:
          content:
            application/json:
              example:
                message: Description of error
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: InternalServerError
      summary: Adds or updates a message.
      tags:
      - message
    put:
      description: Adds or updates a message
      operationId: putMessage
      requestBody:
        content:
          application/json:
            example:
              $ref: examples/body/example-message.json
            schema:
              $ref: '#/components/schemas/Message'
        description: Adds or updates a message
      responses:
        204:
          description: request to update item completed
        400:
          content:
            application/json:
              example:
                message: Bad Request
                errors:
                - message: Value Exceeded. Must be between {min} and {max}.
                  key: common.api.shoesize
                  context:
                    max: "50"
                    min: "1"
              schema:
                $ref: '#/components/schemas/BadRequestError'
          description: BadRequest
        500:
          content:
            application/json:
              example:
                message: Description of error
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: InternalServerError
      summary: Adds or updates a message.
      tags:
      - message
    summary: Message
  /message/messages:
    description: No description available
    get:
      description: Returns all available messages
      operationId: getMessages
      responses:
        200:
          content:
            application/json:
              example:
                $ref: examples/body/example-messages-get.json
              schema:
                $ref: '#/components/schemas/MessagesGetResponseBody'
          description: No description available
        400:
          content:
            application/json:
              example:
                message: Bad Request
                errors:
                - message: Value Exceeded. Must be between {min} and {max}.
                  key: common.api.shoesize
                  context:
                    max: "50"
                    min: "1"
              schema:
                $ref: '#/components/schemas/BadRequestError'
          description: BadRequest
        500:
          content:
            application/json:
              example:
                message: Description of error
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: InternalServerError
      summary: Returns all available messages.
      tags:
      - message
    summary: /messages
components:
  schemas:
    Message:
      $ref: schemas/body/message.json
    MessagesGetResponseBody:
      $ref: schemas/body/messages-get.json
    MessagePostResponseBody:
      $ref: schemas/body/message-created.json
    BadRequestError:
      properties:
        message:
          description: Any further information
          type: string
        errors:
          description: Detailed error information
          items:
            $ref: '#/components/schemas/ErrorItem'
          type: array
      required:
      - message
      type: object
      x-java-type: com.backbase.buildingblocks.presentation.errors.BadRequestException
    ErrorItem:
      description: A validation error
      properties:
        message:
          description: Default Message. Any further information.
          type: string
        key:
          description: '{capability-name}.api.{api-key-name}. For generated validation
            errors this is the path in the document the error resolves to. e.g. object
            name + ''.'' + field'
          type: string
        context:
          description: Context can be anything used to construct localised messages.
          type: object
          x-java-type: java.util.Map<String,String>
      type: object
    ForbiddenError:
      properties:
        message:
          description: Any further information
          type: string
        errors:
          description: Detailed error information
          items:
            $ref: '#/components/schemas/ErrorItem'
          type: array
      type: object
      x-java-type: com.backbase.buildingblocks.presentation.errors.ForbiddenException
    InternalServerError:
      description: Represents HTTP 500 Internal Server Error
      properties:
        message:
          description: Further Information
          type: string
      required:
      - message
      type: object
      x-java-type: com.backbase.buildingblocks.presentation.errors.InternalServerErrorException
    NotAcceptableError:
      properties:
        message:
          type: string
        supportedMediaTypes:
          description: List of supported media types for this endpoint
          items:
            type: string
          type: array
      type: object
      x-java-type: com.backbase.buildingblocks.presentation.errors.NotAcceptableException
    NotFoundError:
      properties:
        message:
          description: Any further information
          type: string
        errors:
          description: Detailed error information
          items:
            $ref: '#/components/schemas/ErrorItem'
          type: array
      type: object
      x-java-type: com.backbase.buildingblocks.presentation.errors.NotFoundException
    UnauthorizedAltError:
      properties:
        message:
          description: Any further information
          type: string
        errors:
          description: Detailed error information
          items:
            $ref: '#/components/schemas/ErrorItem'
          type: array
      type: object
      x-java-type: com.backbase.buildingblocks.presentation.errors.UnauthorizedException
    UnauthorizedError:
      properties:
        message:
          type: string
      required:
      - message
      type: object
    UnsupportedMediaTypeError:
      properties:
        message:
          description: Any further information
          type: string
        errors:
          description: Detailed error information
          items:
            $ref: '#/components/schemas/ErrorItem'
          type: array
      type: object
      x-java-type: com.backbase.buildingblocks.presentation.errors.UnsupportedMediaTypeException